Sort out [Object object] errors
We were return an { error: ... } object from the base query as per rtkq docs, but that stringifies the structured error. Looks like throwing inside the base query has the desired outcome.
