+++ apps/desktop/src/lib/error/parser.ts
@@ -1,6 +1,6 @@
+import { isTauriCommandError } from '$lib/backend/ipc';
 import { KNOWN_ERRORS } from '$lib/error/knownErrors';
 import {
-	isBackendError,
 	isHttpError,
 	isPromiseRejection,
 	isReduxActionError as isReduxActionError
@@ -13,9 +13,12 @@ export interface ParsedError {
 }
 
 export function parseError(error: unknown): ParsedError {
+	if (error instanceof PromiseRejectionEvent && isTauriCommandError(error.reason)) {
+		return { parsedError: error.reason.message };
+	}
 	if (isPromiseRejection(error)) {
 		return { message: 'A promise had an unhandled exception.', parsedError: String(error.reason) };
-	} else if (isBackendError(error) && error.code in KNOWN_ERRORS) {
+	} else if (isTauriCommandError(error) && error.code && error.code in KNOWN_ERRORS) {
 		return { message: KNOWN_ERRORS[error.code], parsedError: error.message };
 	} else if (isReduxActionError(error)) {
 		return { parsedError: error.error + '\n\n' + error.payload };
